predict.py/predict_stream.pyをLogKVLMに対応 (config.jsonのmodel_typeで自動判別)
_load_modelがmodel_type=="logkv"ならLogKVLM.from_pretrainedで読み込む。
REPL本体(TextStreamer/Ctrl+C中断/Alt+Enter改行/サンプリング)は無変更で
そのまま動く。あわせてbf16重み(autocastなし)推論でonline softmaxの
p(fp32昇格)とv_slots(bf16)のeinsumが型不一致になるバグを修正
(pをvalueのdtypeに揃える=autocast時と同じ挙動、累積はfp32のまま)。
bf16重み推論のテストを追加(119件パス)。

+++ logkv.py
@@ -280,7 +280,9 @@ class LogKV(nn.Module):
         m_safe = torch.where(m == float('-inf'), torch.zeros_like(m), m)
         p = torch.exp(logits.to(acc_dtype) - m_safe[..., None])
         l = p.sum(dim=-1)
-        acc = torch.einsum('blc,blcd->bld', p, v_slots).to(acc_dtype)
+        # matmul in the value dtype (what autocast does anyway; also makes
+        # bf16-weight inference without autocast work), accumulate in >= fp32
+        acc = torch.einsum('blc,blcd->bld', p.to(v_slots.dtype), v_slots).to(acc_dtype)
         return m, l, acc
 
     def predict(self, x, hidden=None):
